begining crowd counting 6/20
