+++ main_cnn.py
@@ -1,0 +1,163 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as f
+from torch.utils.data import DataLoader
+from torchvision import datasets, transforms
+import matplotlib.pyplot as plt
+
+class GAP(nn.Module):
+    def __init__(self):
+        super().__init__()
+
+    def forward(self, x):
+        return f.avg_pool2d(x, kernel_size=x.size(2)).view(-1, x.size(1))
+
+class MyNet(torch.nn.Module):
+    def __init__(self):
+        super(MyNet, self).__init__()
+
+        self.conv1 = nn.Conv2d(1,32,3)
+        self.conv2 = nn.Conv2d(32,64,3)
+
+        self.bn1 = nn.BatchNorm2d(32)
+        self.bn2 = nn.BatchNorm2d(64)
+
+        self.pool = nn.MaxPool2d(2, stride=2)
+        self.relu = nn.ReLU()
+
+        self.GAP = GAP()
+
+        self.fc = nn.Linear(64, 10)
+        
+
+    def forward(self, x):
+        x = self.conv1(x)
+        x = self.bn1(x)
+        x = self.relu(x)
+
+        x = self.pool(x)
+
+        x = self.conv2(x)
+        x = self.bn2(x)
+        x = self.relu(x)
+
+        x = self.GAP(x)
+
+        x = self.fc(x)
+
+        return f.log_softmax(x, dim=1)
+
+
+def load_MNIST(batch=128, intensity=1.0):
+    train_loader = torch.utils.data.DataLoader(
+        datasets.MNIST('./data',
+                       train=True,
+                       download=True,
+                       transform=transforms.Compose([
+                           transforms.ToTensor(),
+                           transforms.Lambda(lambda x: x * intensity)
+                       ])),
+        batch_size=batch,
+        shuffle=True)
+
+    test_loader = torch.utils.data.DataLoader(
+        datasets.MNIST('./data',
+                       train=False,
+                       transform=transforms.Compose([
+                           transforms.ToTensor(),
+                           transforms.Lambda(lambda x: x * intensity)
+                       ])),
+        batch_size=batch,
+        shuffle=True)
+
+    return {'train': train_loader, 'test': test_loader}
+
+
+if __name__ == '__main__':
+    # 学習回数
+    epoch = 20
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+    # 学習結果の保存用
+    history = {
+        'train_loss': [],
+        'test_loss': [],
+        'test_acc': [],
+    }
+
+    # ネットワークを構築
+    net: torch.nn.Module = MyNet().to(device)
+    print(net)
+
+    print(sum(p.numel() for p in net.parameters() if p.requires_grad))
+
+    # MNISTのデータローダーを取得
+    loaders = load_MNIST()
+
+    optimizer = torch.optim.Adam(params=net.parameters(), lr=0.001)
+
+    for e in range(epoch):
+
+        """ Training Part"""
+        loss = None
+        # 学習開始 (再開)
+        net.train(True)  # 引数は省略可能
+        for i, (data, target) in enumerate(loaders['train']):
+            # 全結合のみのネットワークでは入力を1次元に
+            # print(data.shape)  # torch.Size([128, 1, 28, 28])
+            # data = data.view(-1, 28*28).to(device)
+            data = data.to(device)
+            target = target.to(device)
+            # print(data.shape)  # torch.Size([128, 784])
+
+            optimizer.zero_grad()
+            output = net(data)
+            loss = f.nll_loss(output, target)
+            loss.backward()
+            optimizer.step()
+
+            if i % 10 == 0:
+                print('Training log: {} epoch ({} / 60000 train. data). Loss: {}'.format(e+1,
+                                                                                         (i+1)*128,
+                                                                                         loss.item())
+                      )
+
+        history['train_loss'].append(loss)
+
+        """ Test Part """
+        # 学習のストップ
+        net.eval()  # または net.train(False) でも良い
+        test_loss = 0
+        correct = 0
+
+        with torch.no_grad():
+            for data, target in loaders['test']:
+                data = data.to(device)
+                target = target.to(device)
+                output = net(data)
+                test_loss += f.nll_loss(output, target, reduction='sum').item()
+                pred = output.argmax(dim=1, keepdim=True)
+                correct += pred.eq(target.view_as(pred)).sum().item()
+
+        test_loss /= 10000
+
+        print('Test loss (avg): {}, Accuracy: {}'.format(test_loss,
+                                                         correct / 10000))
+
+        history['test_loss'].append(test_loss)
+        history['test_acc'].append(correct / 10000)
+
+    # 結果の出力と描画
+    print(history)
+    plt.figure()
+    plt.plot(range(1, epoch+1), history['train_loss'], label='train_loss')
+    plt.plot(range(1, epoch+1), history['test_loss'], label='test_loss')
+    plt.xlabel('epoch')
+    plt.legend()
+    plt.savefig('results/loss_cnn.png')
+
+    plt.figure()
+    plt.plot(range(1, epoch+1), history['test_acc'])
+    plt.title('test accuracy')
+    plt.xlabel('epoch')
+    plt.savefig('results/test_acc_cnn.png')
