+++ results/loss_cnn.png
